Use explicit NEWLINE_STYLE for version.hpp
The file include/version.hpp is configured by CMake and also commited
to the repo.  CMake will generate this file with different line endings
depending on the platform, however, the file commited to the repo uses
LF line endings.  This PR modifies the CMake to always generates LF
line endings so its output always matches what has been commited to
the repo.

I've also added ".gitattributes" to make sure that git doesn't modify
the line endings of any ".hpp" depending on the users "core.autocrlf"
setting.  Without this, there would still be a discrepancy to address
depending on whether this setting is enabled.